Refactor move ShoeDTO to parent package
