+++ dto/src/main/java/com/example/demo/dto/IShoe.java
@@ -1,0 +1,4 @@
+package com.example.demo.dto;
+
+public interface IShoe {
+}
