+++ dto/src/main/java/com/example/demo/dto/ShoeV3.java
@@ -1,4 +1,4 @@
-package com.example.demo.dto.out;
+package com.example.demo.dto;
 
 import lombok.Builder;
 import lombok.Value;
